Use ESM only and move link from utils so its exported properly
